requirements.txt
PyYAML==6.0
python-box==6.0.2
pyTelegramBotAPI==4.6.0
gspread==5.4.0
schedule==1.1.0
matplotlib==3.5.2
nats-py==2.1.4
aiohttp==3.8.1
